Updated SQUAD (#427)
* Metalands Dapp
Please review, And Pull Request to show Metalands V Dapp in App Hub.

* Update promoted.ts

* Delete logo.png

* Add files via upload

* Update manifest.json

* BVM Logo Update + Priority List

A small update on the Logo of BlackVeMarket, Aswell as the description.

Also a request to be on the promoted links on app hub.

* Delete apps/com.metalandsv directory

* Metalands Dapp

Pull request to merge MetaLands Mint page into App Hub

* Delete

* Update promoted.ts

* Metalands Dapp

Request to add Metalands to App-Hub

* SQUAD

* Update manifest.json

* Turtle Swap Commit

* not relevant

* fgzg

* BangzBoardz Commit

* commit Bangz

* PB Commit

* GE

* Update manifest.json

* Dreamchick commit

* Clayworld commit

* PR for update Squad Vechain URL

* remove.

* Add new URL SQUAD

---------

Co-authored-by: libotony <liboliqi@gmail.com>